On branch master Your branch is up-to-date with 'origin/master'. Changes to be committed:   (use "git reset HEAD <file>..." to unstage)
	modified:   xdg/config.sh

+++ xdg/config.sh
@@ -33,11 +33,13 @@ auto_distroicon() {
 	if [ -f /etc/manjaro-release ];then
 		_distributor="manjaro"
 	fi
-	for _dir in $(echo $(find -maxdepth 1 -mindepth 1 -type d));do
-		cd $_dir
-		cp -fv logos/emblem-$_distributor.png logos/emblem-distributor.png
-		cd $_basedir
-	done
+	if [ ! -f 48/logos/emblem-$_distributor.png ];then
+		for _dir in $(echo $(find -maxdepth 1 -mindepth 1 -type d));do
+			cd $_dir
+			cp -fv logos/emblem-$_distributor.png logos/emblem-distributor.png
+			cd $_basedir
+		done
+	fi
 }
 
 custom_distroicon() {
